Split tests into two files
